Rename `p6_code_analysis` to `p7_code_analysis`
