+++ crates/solidity-v2/outputs/cargo/semantic/src/context/mod.rs
@@ -11,7 +11,7 @@ use slang_solidity_v2_ir::ir;
 use crate::binder::{Binder, Definition, Reference};
 use crate::passes::{
     p1_collect_definitions, p2_linearise_contracts, p3_type_definitions, p4_compute_linearisations,
-    p5_resolve_references, p6_code_analysis, p6_resolve_yul,
+    p5_resolve_references, p6_resolve_yul, p7_code_analysis,
 };
 use crate::types::{
     ArrayType, ByteArrayType, ContractType, EnumType, FixedPointNumberType, FixedSizeArrayType,
@@ -83,7 +83,7 @@ impl SemanticContext {
         let contract_data = p4_compute_linearisations::run(&binder, &types);
         p5_resolve_references::run(files, &mut binder, &mut types);
         p6_resolve_yul::run(files, &mut binder, &types, diagnostics);
-        p6_code_analysis::run(
+        p7_code_analysis::run(
             &binder,
             language_version,
             evm_target,
