+++ crates/solidity-v2/outputs/cargo/semantic/src/passes/mod.rs
@@ -4,8 +4,8 @@ pub(crate) mod p2_linearise_contracts;
 pub(crate) mod p3_type_definitions;
 pub(crate) mod p4_compute_linearisations;
 pub(crate) mod p5_resolve_references;
-pub(crate) mod p6_code_analysis;
 pub(crate) mod p6_resolve_yul;
+pub(crate) mod p7_code_analysis;
 
 #[cfg(test)]
 mod tests;
