+++ crates/solidity-v2/outputs/cargo/semantic/src/passes/p7_code_analysis/built_ins/validate_built_ins.generated.rs
